# src/GuildSaber.Website/openapi/GuildSaber.Api.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "GuildSaber.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/ranked-maps/{rankedMapId}": {
      "get": {
        "tags": [
          "RankedMaps"
        ],
        "summary": "Get a ranked map.",
        "description": "Get a ranked map in the server by its Id.",
        "operationId": "GetRankedMap",
        "parameters": [
          {
            "name": "rankedMapId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankedMap"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.5",
                  "title": "Not Found",
                  "status": 404,
                  "instance": "GET ranked-maps/{rankedMapId}",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          }
        }
      }
    },
    "/ranked-maps/{rankedMapId}/with-scores/{playerId}": {
      "get": {
        "tags": [
          "RankedMaps"
        ],
        "summary": "Get a ranked map with the player's point scores.",
        "description": "Get a ranked map in the server by its Id, including the player's best point scores on the map.",
        "operationId": "GetRankedMapWithScores",
        "parameters": [
          {
            "name": "rankedMapId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankedMapWithScores"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.5",
                  "title": "Not Found",
                  "status": 404,
                  "instance": "GET ranked-maps/{rankedMapId}/with-scores/{playerId}",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          }
        }
      }
    },
    "/ranked-maps/{rankedMapId}/with-scores/@me": {
      "get": {
        "tags": [
          "RankedMaps"
        ],
        "summary": "Get a ranked map with the current player's point scores.",
        "description": "Get a ranked map in the server by its Id, including the current player's best point scores on the map.",
        "operationId": "GetRankedMapWithScoresAtMe",
        "parameters": [
          {
            "name": "rankedMapId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankedMapWithScores"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.5",
                  "title": "Not Found",
                  "status": 404,
                  "instance": "GET ranked-maps/{rankedMapId}/with-scores/@me",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ]
          }
        ]
      }
    },
...
    "/contexts/{contextId}/ranked-maps/with-scores/{playerId}": {
      "get": {
        "tags": [
          "Context.RankedMaps"
        ],
        "summary": "Get ranked maps for a context with the player's point scores.",
        "description": "Get ranked maps for a context by its Id, with optional search and sorting, including the player's best point scores on each map.",
        "operationId": "GetRankedMapsWithScores",
        "parameters": [
          {
            "name": "contextId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ContextId"
            }
          },
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PlayerId"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "pattern": "^-?(?:0|[1-9]\\d*)$",
                "type": [
                  "integer",
                  "string"
                ],
                "format": "int32"
              }
            }
          },
          {
            "name": "matchAnyCategory",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "rankedScoreTypes",
            "in": "query",
            "schema": {
              "default": "None",
              "$ref": "#/components/schemas/ERankedScoreType"
            }
          },
          {
            "name": "includeMapsWithoutScore",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "difficultyStarFrom",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "float"
            }
          },
          {
            "name": "difficultyStarTo",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "float"
            }
          },
          {
            "name": "accuracyStarFrom",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "float"
            }
          },
          {
            "name": "accuracyStarTo",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "float"
            }
          },
          {
            "name": "durationSecFrom",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "float"
            }
          },
          {
            "name": "durationSecTo",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "float"
            }
          },
          {
            "name": "bpmFrom",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "float"
            }
          },
          {
            "name": "bpmTo",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "float"
            }
          },
          {
            "name": "needConfirmation",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "default": "DifficultyStar",
              "$ref": "#/components/schemas/ERankedMapSorter"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "default": "Asc",
              "$ref": "#/components/schemas/EOrder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedListOfRankedMapWithScores"
                }
              }
            }
          }
        }
      }
    },
...
    "/contexts/{contextId}/ranked-maps/{rankedMapId}": {
      "put": {
        "tags": [
          "Context.RankedMaps"
        ],
        "summary": "Update a ranked map.",
        "description": "Update a ranked map by its Id.",
        "operationId": "UpdateRankedMap",
        "parameters": [
          {
            "name": "contextId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rankedMapId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRankedMap"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankedMapSimple"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                  "title": "Bad Request",
                  "status": 400,
                  "detail": null,
                  "instance": "PUT contexts/{contextId}/ranked-maps/{rankedMapId}",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.5",
                  "title": "Not Found",
                  "status": 404,
                  "detail": null,
                  "instance": "PUT contexts/{contextId}/ranked-maps/{rankedMapId}",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContextId": {
        "type": "integer",
        "format": "int32"
      },
...
      "PlayerId": {
        "type": "integer",
        "format": "int32"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "RankedMap": {
        "required": [
          "id",
          "guildId",
          "contextId",
          "info",
          "requirements",
          "rating",
          "versions",
          "categoryIds",
          "achievements"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RankedMapId"
          },
          "guildId": {
            "$ref": "#/components/schemas/GuildId"
          },
          "contextId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "info": {
            "$ref": "#/components/schemas/RankedMapInfo"
          },
          "requirements": {
            "$ref": "#/components/schemas/RankedMapRequirements"
          },
          "rating": {
            "$ref": "#/components/schemas/RankedMapRating"
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MapVersion"
            }
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          "achievements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Achievement"
            }
          }
        }
      },
...
      "RankedMapWithScores": {
        "required": [
          "rankedMap",
          "rankedScores"
        ],
        "type": "object",
        "properties": {
          "rankedMap": {
            "$ref": "#/components/schemas/RankedMap"
          },
          "rankedScores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RankedScore"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "description": "GuildSaber session cookie. Browsers and cookie-enabled HTTP clients send it automatically.",
        "name": "__Host-GuildSaber.Session",
        "in": "cookie"
      },
      "X-GuildSaber-Request": {
        "type": "apiKey",
        "description": "Required for state-changing requests authenticated with the session cookie.",
        "name": "X-GuildSaber-Request",
        "in": "header"
      }
    }
  },
...
}